Refactor into module and add comments
